+++ src/components/Roster/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import { StyledRoster } from './Styled';
+import { BaseProps } from '../Base';
+
+export const Roster: React.FC<BaseProps> = ({ children, ...rest }) => {
+  return <StyledRoster {...rest}>{children}</StyledRoster>;
+};
+
+export default Roster;