+++ CHANGELOG.md
@@ -8,14 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-Add Play, Pause and Sound disabled SVG icons
+
+- Add Play, Pause and Sound disabled SVG icons
+- Add initial roster components
 
 ### Changed
+
 Update prebuild script to trigger git actions on git push event
 
 ### Removed
 
 ### Fixed
+
 Fixed prebuild for PR and Push.
 
 ## [0.1.1] - 2020-06-16